Rest use long long for card_id.

+++ QT_frontend/DLLRestAPI/dll_rest_api.h
@@ -25,7 +25,7 @@ class DLLRESTAPI_EXPORT dll_rest_api : public QObject {
 public:
     dll_rest_api(QObject *parent = nullptr);
     ~dll_rest_api();
-    void login(QByteArray hash, int card_num);
+    void login(QByteArray hash, long long card_num);
     void get_customer_info(int customer_id);
     void get_account_info(int account_id);
     void get_card_status(int card_num);
